requirements.txt
line-bot-sdk
flask
psycopg2  #需要postgres時，請雲端幫我裝
